+++ packages/marketo-form/partials/field-wrapper/index.tsx
@@ -5,15 +5,18 @@ import styles from './style.module.css'
 const FieldWrapper = ({
   size,
   children,
+  fieldId,
 }: {
   size?: 'lg'
   children: ReactNode
+  fieldId: string
 }) => {
   return (
     <div
       className={clsx('marketo-form-field-wrapper', styles.wrapper, {
         [styles.sizeLg]: size === 'lg',
       })}
+      data-field-id={fieldId}
     >
       {children}
     </div>